Update type Webpage to remove domain import
